+++ app/blogs/components/SearchBar.tsx
@@ -1,0 +1,24 @@
+'use client'
+
+import { Input } from '@/components/ui/input'
+import { Search } from 'lucide-react'
+
+interface SearchBarProps {
+    searchQuery: string;
+    setSearchQuery: (query: string) => void;
+}
+
+export default function SearchBar({ searchQuery, setSearchQuery }: SearchBarProps) {
+    return (
+        <div className="relative mb-8">
+            <Search className="absolute left-3 top-1/2 transform -translate-y-1/2 text-muted-foreground h-4 w-4" />
+            <Input
+                type="text"
+                placeholder="Search blog post content..."
+                value={searchQuery}
+                onChange={(e) => setSearchQuery(e.target.value)}
+                className="pl-10"
+            />
+        </div>
+    )
+}
