--- app/home/layout.tsx
@@ -1,12 +1,0 @@
-export const metadata = {
-  title: 'Home',
-  description: 'Jake Bodea\'s home page'
-}
-
-export default function HomeLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return children;
-}
