+++ app/projects/components/ProjectHeader.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const ProjectHeader = () => (
+    <div className="mb-12">
+        <h1 className="text-4xl md:text-5xl font-serif font-light text-foreground mb-4">
+            Featured Projects
+        </h1>
+        <p className="text-xl text-muted-foreground font-light leading-relaxed max-w-2xl">
+            A showcase of my technical work, from AI applications to full-stack solutions.
+        </p>
+    </div>
+);
+
+export default ProjectHeader;
