+++ app/timeline/components/TimelineHeader.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const TimelineHeader = () => (
+  <div className="mb-12">
+    <h1 className="text-4xl md:text-5xl font-serif font-light text-foreground mb-4">
+      Professional Timeline
+    </h1>
+    <p className="text-xl text-muted-foreground font-light leading-relaxed max-w-2xl">
+      A chronological journey through my career experiences, from student to AI engineer.
+    </p>
+  </div>
+);
+
+export default TimelineHeader;
