+++ components/layout/Sidebar.tsx
@@ -3,10 +3,10 @@
 import dynamic from 'next/dynamic';
 
 const AppSidebar = dynamic(
-  () => import('@/components/app-sidebar').then(mod => mod.AppSidebar),
+  () => import('@/components/layout/app-sidebar').then(mod => mod.AppSidebar),
   { ssr: false }
 );
 
-export function SidebarWrapper() {
+export function Sidebar() {
   return <AppSidebar />;
 }
