+++ content/sidebar.ts
@@ -1,0 +1,41 @@
+export const sidebarData = {
+  pages: [
+    {
+      title: "Home",
+      url: "/",
+      icon: "🏠"
+    },
+    {
+      title: "Timeline",
+      url: "/timeline",
+      icon: "⏳"
+    },
+    {
+      title: "Projects",
+      url: "/projects",
+      icon: "💻"
+    },
+    {
+      title: "Blogs",
+      url: "/blogs",
+      icon: "📝"
+    }
+  ],
+  contact: [
+    {
+      title: "LinkedIn",
+      url: "https://www.linkedin.com/in/jakebodea/",
+      icon: "🖇"
+    },
+    {
+      title: "GitHub",
+      url: "https://github.com/jakebodea",
+      icon: "🐙"
+    },
+    {
+      title: "𝕏",
+      url: "https://x.com/jakebodea",
+      icon: "🐦"
+    }
+  ]
+};
